TweetDeck unofficial : Icon proposal to fit with the uTouch graphical chart
Hi Shinsuke,

I'm an amateur french graphic designer (with GIMP of course ;)) and I've an Ubuntu phone since a few months. As I love the community's spirit in this OS, I try to look after the Ubuntu store and propose my help to those who want it.

Now I just download and use your webapp and I love it. But as this is an unofficial application, I thought we could have a little fun with the logo and I propose this one (kind of flat design) in two sizes : 256 x 256 and 512 x 512 pixels by side.

Tell me if you're interested and have a good day in all cases !

Best regards,
Franck